+++ beets/library.py
@@ -19,7 +19,6 @@ from __future__ import (division, absolute_import, print_function,
 
 import os
 import sys
-import shlex
 import unicodedata
 import time
 import re
@@ -1139,13 +1138,8 @@ def parse_query_string(s, model_cls):
     The string is split into components using shell-like syntax.
     """
     assert isinstance(s, unicode), "Query is not unicode: {0!r}".format(s)
-
-    # A bug in Python < 2.7.3 prevents correct shlex splitting of
-    # Unicode strings.
-    # http://bugs.python.org/issue6988
-    s = s.encode('utf8')
     try:
-        parts = [p.decode('utf8') for p in shlex.split(s)]
+        parts = util.shlex_split(s)
     except ValueError as exc:
         raise dbcore.InvalidQueryError(s, exc)
     return parse_query_parts(parts, model_cls)