+++ test/test_util.py
@@ -43,7 +43,7 @@ class UtilTest(unittest.TestCase):
     @patch('beets.util.open_anything')
     def test_interactive_open(self, mock_open, mock_execlp):
         mock_open.return_value = 'tagada'
-        util.interactive_open(['foo'])
+        util.interactive_open(['foo'], util.open_anything())
         mock_execlp.assert_called_once_with('tagada', 'tagada', 'foo')
         mock_execlp.reset_mock()